ire scusa stavo provando a da na sistemata ai commenti, a livell teorico dovrebbe funziona, a livello pratico no, da un errore molto strano ricordati di cambiare il database ho cambiato i campi del commento! il resto lo trovi nei commenti del codice committato

+++ src/main/java/it/uniroma3/siw/poesia/siwpoesia0/service/ImmagineService.java
@@ -32,7 +32,8 @@ public class ImmagineService {
     }
 
     public void deleteImmagine(Immagine img) {
-       this.immagineRepository.delete(img);
+        if(img == null) return;
+        this.immagineRepository.delete(img);
     }
 
     public Immagine find(Long id) {
